+++ sdk/communication/communication-administration/recordings/node/communicationidentityclient_playbacklive/recording_successfully_deletes_a_user.js
@@ -1,6 +1,6 @@
 let nock = require("nock");
 
-module.exports.hash = "c32f7ebc5af64d902e3e080cc228e00f";
+module.exports.hash = "5f3b4220446ac81ca7796c4e29d977c5";
 
 module.exports.testInfo = { uniqueName: {}, newDate: {} };
 
@@ -9,7 +9,7 @@ nock("https://endpoint", { encodedQueryParams: true })
   .query(true)
   .reply(204, "", [
     "MS-CV",
-    "62IvL0rpM0WTit4IOMyxyA.0",
+    "HHCDHFV8yEySqM6/Ua+73w.0",
     "Strict-Transport-Security",
     "max-age=2592000",
     "x-ms-client-request-id",
@@ -17,9 +17,9 @@ nock("https://endpoint", { encodedQueryParams: true })
     "api-supported-versions",
     "2020-07-20-preview2, 2021-03-07",
     "X-Processing-Time",
-    "180ms",
+    "178ms",
     "X-Azure-Ref",
-    "0JP0GYAAAAAAWVrFMxKQDT5rgGt1tHUQSWVZSMzBFREdFMDMxOAA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+    "0W+EJYAAAAAB89UA1+UkvQoORF0b/3woTV1NURURHRTA4MTcAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
     "Date",
-    "Tue, 19 Jan 2021 15:39:16 GMT"
+    "Thu, 21 Jan 2021 20:17:32 GMT"
   ]);