+++ sdk/communication/communication-administration/recordings/node/communicationidentityclient_playbacklive/recording_successfully_issues_a_token_for_a_user_multiple_scopes.js
@@ -1,19 +1,19 @@
 let nock = require("nock");
 
-module.exports.hash = "bd62b879c364ef58cd06ba00781e7d17";
+module.exports.hash = "4a6aa78ff3e608a0c86d94e8211ad5bc";
 
 module.exports.testInfo = { uniqueName: {}, newDate: {} };
 
 nock("https://endpoint", { encodedQueryParams: true })
   .post("/identities/sanitized/:issueAccessToken", { scopes: ["chat", "voip"] })
   .query(true)
-  .reply(200, { token: "sanitized", expiresOn: "2021-01-20T15:39:14.7969446+00:00" }, [
+  .reply(200, { token: "sanitized", expiresOn: "2021-01-22T20:17:30.1353667+00:00" }, [
     "Transfer-Encoding",
     "chunked",
     "Content-Type",
     "application/json; charset=utf-8",
     "MS-CV",
-    "sxVVnmTVpUWSRiMteR9pnQ.0",
+    "FcCTQ857p0Kmt16FfR/9cw.0",
     "Strict-Transport-Security",
     "max-age=2592000",
     "x-ms-client-request-id",
@@ -21,9 +21,9 @@ nock("https://endpoint", { encodedQueryParams: true })
     "api-supported-versions",
     "2020-07-20-preview2, 2021-03-07",
     "X-Processing-Time",
-    "28ms",
+    "26ms",
     "X-Azure-Ref",
-    "0I/0GYAAAAADVZIvEUo8zRrHkN6cf+BwGWVZSMzBFREdFMDMxOAA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+    "0W+EJYAAAAADqzza+TSriRrm7vmstTyXWV1NURURHRTA4MTcAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
     "Date",
-    "Tue, 19 Jan 2021 15:39:15 GMT"
+    "Thu, 21 Jan 2021 20:17:31 GMT"
   ]);