+++ sdk/communication/communication-administration/recordings/node/communicationidentityclient_playbacklive/recording_successfully_issues_a_token_for_a_user_single_scope.js
@@ -1,19 +1,19 @@
 let nock = require("nock");
 
-module.exports.hash = "1c880cc35f2f3b240445a35a012e7bab";
+module.exports.hash = "c6f3ddd7c26e3db3fdb740807d8806c6";
 
 module.exports.testInfo = { uniqueName: {}, newDate: {} };
 
 nock("https://endpoint", { encodedQueryParams: true })
   .post("/identities/sanitized/:issueAccessToken", { scopes: ["chat"] })
   .query(true)
-  .reply(200, { token: "sanitized", expiresOn: "2021-01-20T15:39:14.7360153+00:00" }, [
+  .reply(200, { token: "sanitized", expiresOn: "2021-01-22T20:17:30.0481704+00:00" }, [
     "Transfer-Encoding",
     "chunked",
     "Content-Type",
     "application/json; charset=utf-8",
     "MS-CV",
-    "Moedc5KSB0eBtdD2heqKCA.0",
+    "10R9kZzcWUOBD+kRuAq4Lg.0",
     "Strict-Transport-Security",
     "max-age=2592000",
     "x-ms-client-request-id",
@@ -21,9 +21,9 @@ nock("https://endpoint", { encodedQueryParams: true })
     "api-supported-versions",
     "2020-07-20-preview2, 2021-03-07",
     "X-Processing-Time",
-    "49ms",
+    "289ms",
     "X-Azure-Ref",
-    "0I/0GYAAAAADymHbD3JAIRauSg01WvZYWWVZSMzBFREdFMDMxOAA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+    "0WuEJYAAAAAAW5X+4jx4JQblAkbrO7KJjV1NURURHRTA4MTcAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
     "Date",
-    "Tue, 19 Jan 2021 15:39:15 GMT"
+    "Thu, 21 Jan 2021 20:17:31 GMT"
   ]);