+++ sdk/communication/communication-administration/CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Added
 
 - `CommunicationIdentityClient` added a constructor that supports `TokenCredential`.
+- `CommunicationIdentityClient` added a new method `createUserWithToken`.
 
 ## 1.0.0-beta.3 (2020-11-16)